requirements.txt
# VK API для работы с ВКонтакте
vk-api>=11.9.9

# Загрузка переменных окружения из .env файла
python-dotenv>=1.0.0

# Google Sheets API для работы с таблицами
gspread>=5.12.0
oauth2client>=4.1.3

# Telegram Bot API для отправки уведомлений
python-telegram-bot>=20.0

# Генерация графиков и диаграмм
matplotlib>=3.8.0

# Обработка и анализ данных
pandas>=2.1.0

# Планировщик задач для автоматизации
APScheduler>=3.10.4

# Опциональные зависимости для цветного вывода в консоль (Windows)
colorama>=0.4.6
